requirements.txt
fastapi==0.115.8
uvicorn[standard]==0.34.0

SQLAlchemy==2.0.37
psycopg[binary]==3.2.4

pgvector==0.3.6

pydantic==2.10.6
pydantic-settings==2.7.1

python-dotenv==1.0.1

openai==1.61.0

pypdf==5.2.0
python-docx==1.1.2